dags/requirements.txt
airflow==2.1.1
python-dateutil==2.8.1
pandas==1.1.4
requests==2.22.0
pyspark==3.1.2
boto3==1.17.84